Revert "Merge pull request #57 from angelod1as/other-table"
This reverts commit 0833abc1e6865536e3fc2d83397200af315a1359, reversing
changes made to 4843e08627aa1f23816ac252f916bd57632ad411.
