+++ app/app.css
@@ -1,12 +1,6 @@
 @import "tailwindcss";
 @import "tw-animate-css";
 
-/* For PrimeReact DataTable */
-@layer primereact;
-@import "primereact/resources/themes/lara-light-indigo/theme.css"
-  layer(primereact);
-/* End PrimeReact DataTable */
-
 @custom-variant dark (&:is(.dark *));
 
 @theme {
