+++ app/components/organisms/data-table/number-box.tsx
