+++ app/components/organisms/event-card/event-card.tsx
@@ -1,5 +1,9 @@
+import { DialogClose } from "@radix-ui/react-dialog"
 import type { FC } from "react"
+import { useFetcher } from "react-router"
+import { Button } from "~/components/atoms/button/button"
 import { DataPair } from "~/components/atoms/data-pair/data-pair"
+import ConfirmDialog from "~/components/molecules/confirm-dialog/confirm-dialog"
 import {
   Card,
   CardContent,
@@ -8,19 +12,42 @@ import {
   CardHeader,
   CardTitle,
 } from "~/components/ui/card"
-
+import {
+  Dialog,
+  DialogContent,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+  DialogTrigger,
+} from "~/components/ui/dialog"
 import { formatDateTime } from "~/lib/helpers/format-date-time"
 import { generateGoogleCalendarLink } from "~/lib/helpers/generate-google-calendar-link"
+import paths from "~/lib/paths"
 import type { ViewEvent } from "~types/entities.types"
 import { DateStatus } from "./date-status"
-import { EventCardFooter } from "./event-card-footer"
+
+const {
+  dash: {
+    participant: { DOWNLOAD_CALENDAR },
+    events: { EVENT_VIEW },
+  },
+} = paths
 
 type EventCardProps = { event: ViewEvent; "data-testid": string }
 export const EventCard: FC<EventCardProps> = ({
   event,
   "data-testid": dataTestId,
 }) => {
-  const googleLink = generateGoogleCalendarLink(event)
+  const fetcher = useFetcher()
+
+  /* Handle Cancel Application */
+  const handleConfirm = async (closeDialog: () => void) => {
+    await fetcher.submit(
+      { cancel: true, eventId: event.id },
+      { method: "POST" },
+    )
+    closeDialog()
+  }
 
   const {
     time_application_end,
@@ -42,6 +69,9 @@ export const EventCard: FC<EventCardProps> = ({
     title,
     is_applied,
   } = event
+
+  const isEventOpen = event_status === "Registration Open"
+  const googleLink = generateGoogleCalendarLink(event)
 
   return (
     <Card className="flex flex-col flex-1" data-testid={dataTestId}>
@@ -84,13 +114,69 @@ export const EventCard: FC<EventCardProps> = ({
         </div>
       </CardContent>
       <CardFooter>
-        <EventCardFooter
-          eventId={id}
-          event_status={event_status}
-          googleLink={googleLink}
-          is_applied={is_applied}
-          dataTestId={dataTestId}
-        />
+        {is_applied ? (
+          <div className="flex flex-col sm:flex-row gap-4 w-full">
+            <Dialog>
+              <DialogTrigger asChild>
+                <Button variant="outline">Adicionar ao Calendário</Button>
+              </DialogTrigger>
+              <DialogContent className="max-w-sm sm:max-w-sm">
+                <DialogHeader>
+                  <DialogTitle>Adicionar ao calendário</DialogTitle>
+                </DialogHeader>
+                <DialogFooter className="flex flex-col sm:justify-between">
+                  {googleLink && (
+                    <DialogClose asChild>
+                      <Button
+                        variant="outline"
+                        to={googleLink}
+                        linkProps={{ target: "_blank" }}
+                      >
+                        Google Calendar
+                      </Button>
+                    </DialogClose>
+                  )}
+                  <DialogClose asChild>
+                    <Button
+                      variant="outline"
+                      to={DOWNLOAD_CALENDAR(event.id)}
+                      linkProps={{ reloadDocument: true }}
+                    >
+                      Baixar arquivo iCal
+                    </Button>
+                  </DialogClose>
+                </DialogFooter>
+              </DialogContent>
+            </Dialog>
+
+            <fetcher.Form method="post">
+              <ConfirmDialog
+                title="Cancelar inscrição"
+                description={
+                  <div>
+                    <p>Você tem certeza que deseja cancelar sua inscrição?</p>
+                  </div>
+                }
+                confirmLabel="😢 Cancelar"
+                cancelLabel="🎉 Voltar"
+                isLoading={fetcher.state !== "idle"}
+                onConfirm={handleConfirm}
+              >
+                <ConfirmDialog.Trigger variant="destructive" className="w-full">
+                  Cancelar inscrição
+                </ConfirmDialog.Trigger>
+              </ConfirmDialog>
+            </fetcher.Form>
+          </div>
+        ) : (
+          <Button
+            data-testid={dataTestId}
+            to={isEventOpen ? EVENT_VIEW(id) : ""}
+            disabled={!isEventOpen}
+          >
+            {isEventOpen ? "Fazer inscrição" : "Inscreva-se em breve"}
+          </Button>
+        )}
       </CardFooter>
     </Card>
   )
