+++ app/pages/admin/events/event-participants-table.tsx/admin-event-participants-table-columns.tsx
@@ -7,12 +7,12 @@ import { EyeIcon } from "lucide-react"
 import WhatsappIcon from "~/assets/social/whatsapp.svg"
 import type { ParticipantWithExtraData } from "~/business/admin/admin.server"
 import { Button } from "~/components/atoms/button/button"
-import { DataTableColumnHeader } from "~/components/organisms/old-data-table/column-header"
-import { selectionBox } from "~/components/organisms/old-data-table/selection-box"
+import { DataTableColumnHeader } from "~/components/organisms/data-table/column-header"
+import { selectionBox } from "~/components/organisms/data-table/selection-box"
 import { phoneToWhatsappLink } from "~/lib/helpers/phone-to-whatsapp-link"
 
-import { makeInput } from "~/components/organisms/old-data-table/make-input"
-import { numberBox } from "~/components/organisms/old-data-table/number-box"
+import { makeInput } from "~/components/organisms/data-table/make-input"
+import { numberBox } from "~/components/organisms/data-table/number-box"
 import { participantProcessStatusPropMap } from "~/lib/helpers/propMaps"
 import paths from "~/lib/paths"
 import type { TableMeta } from "~/types/table.types"
