+++ app/pages/admin/events/view-event.tsx
@@ -4,7 +4,7 @@ import { redirectWithError } from "remix-toast"
 import { toast } from "sonner"
 import {
   getAdminContext,
-  getSupabaseAdminEventById,
+  getAdminEventById,
   updateEventStatus,
 } from "~/business/admin/admin.server"
 import { updateEventStatusSchema } from "~/business/admin/common"
@@ -34,7 +34,7 @@ export async function action({ request, params }: Route.ActionArgs) {
 }
 
 export async function loader({ params, request }: Route.LoaderArgs) {
-  const event = await getSupabaseAdminEventById(request, params)
+  const event = await getAdminEventById(request, params)
   if (!event) {
     throw await redirectWithError(ADMIN_DASHBOARD, "Evento não encontrado")
   }
