+++ app/pages/dashboard/dashboard-page.tsx
@@ -16,14 +16,9 @@ const {
 } = paths
 
 const splitEvents = (events: ViewEvent[] | undefined) => {
-  const empty: {
-    registrationOpen: ViewEvent[]
-    scheduled: ViewEvent[]
-    registrationClosed: ViewEvent[]
-  } = {
+  const empty: { registrationOpen: ViewEvent[]; scheduled: ViewEvent[] } = {
     registrationOpen: [],
     scheduled: [],
-    registrationClosed: [],
   }
 
   if (!events || events.length < 1) return empty
@@ -31,8 +26,6 @@ const splitEvents = (events: ViewEvent[] | undefined) => {
   return events.reduce((acc, event) => {
     if (event.event_status === "Registration Open") {
       acc.registrationOpen.push(event)
-    } else if (event.event_status === "Registration Closed") {
-      acc.registrationClosed.push(event)
     } else {
       acc.scheduled.push(event)
     }
@@ -56,8 +49,7 @@ export async function clientLoader({}: Route.LoaderArgs) {
   if (error || !events) {
     return {
       registrationOpen: [],
-      scheduled: undefined,
-      registrationClosed: undefined,
+      scheduled: [],
     }
   }
 
@@ -69,42 +61,26 @@ export async function action({ request, params }: Route.ClientActionArgs) {
 }
 
 type WrapperProps = {
-  openRegistrationEvents: ReactNode
+  registrationEvents: ReactNode
   scheduledEvents: ReactNode
-  closedRegistrationEvents: ReactNode
 }
 
-const Wrapper: FC<WrapperProps> = ({
-  openRegistrationEvents,
-  scheduledEvents,
-  closedRegistrationEvents,
-}) => {
+const Wrapper: FC<WrapperProps> = ({ registrationEvents, scheduledEvents }) => {
   return (
     <>
       <div className="flex flex-col gap-4">
         <h2>Inscrições abertas</h2>
         <div className="grid gap-4 grid-cols-1 lg:grid-cols-2 2xl:grid-cols-3">
-          {openRegistrationEvents}
+          {registrationEvents}
         </div>
       </div>
 
-      {closedRegistrationEvents && (
-        <div className="flex flex-col gap-4">
-          <h2>Inscrições encerradas</h2>
-          <div className="grid gap-4 grid-cols-1 lg:grid-cols-2 2xl:grid-cols-3">
-            {closedRegistrationEvents}
-          </div>
+      <div className="flex flex-col gap-4">
+        <h2>Eventos agendados</h2>
+        <div className="grid gap-4 grid-cols-1 lg:grid-cols-2 2xl:grid-cols-3">
+          {scheduledEvents}
         </div>
-      )}
-
-      {scheduledEvents && (
-        <div className="flex flex-col gap-4">
-          <h2>Eventos agendados</h2>
-          <div className="grid gap-4 grid-cols-1 lg:grid-cols-2 2xl:grid-cols-3">
-            {scheduledEvents}
-          </div>
-        </div>
-      )}
+      </div>
     </>
   )
 }
@@ -112,19 +88,18 @@ const Wrapper: FC<WrapperProps> = ({
 export const HydrateFallback = () => {
   return (
     <Wrapper
-      openRegistrationEvents={<EventCardSkeleton />}
-      scheduledEvents={undefined}
-      closedRegistrationEvents={undefined}
+      registrationEvents={<EventCardSkeleton />}
+      scheduledEvents={<EventCardSkeleton />}
     />
   )
 }
 
 const DashboardPage = ({ loaderData }: Route.ComponentProps) => {
-  const { registrationOpen, registrationClosed, scheduled } = loaderData
+  const { registrationOpen, scheduled } = loaderData
 
   return (
     <Wrapper
-      openRegistrationEvents={
+      registrationEvents={
         registrationOpen.length ? (
           registrationOpen.map((event) => (
             <EventCard
@@ -137,25 +112,18 @@ const DashboardPage = ({ loaderData }: Route.ComponentProps) => {
           <p>Nenhum evento com inscrições abertas</p>
         )
       }
-      closedRegistrationEvents={
-        !!registrationClosed?.length &&
-        registrationClosed.map((event) => (
-          <EventCard
-            data-testid="event-card-closed"
-            key={event.id}
-            event={event}
-          />
-        ))
-      }
       scheduledEvents={
-        !!scheduled?.length &&
-        scheduled.map((event) => (
-          <EventCard
-            data-testid="event-card-scheduled"
-            key={event.id}
-            event={event}
-          />
-        ))
+        scheduled.length ? (
+          scheduled.map((event) => (
+            <EventCard
+              data-testid="event-card-scheduled"
+              key={event.id}
+              event={event}
+            />
+          ))
+        ) : (
+          <p>Nenhum evento encontrado</p>
+        )
       }
     />
   )
