+++ package.json
@@ -54,7 +54,6 @@
     "nodemailer": "^7.0.3",
     "pg": "^8.16.0",
     "prettier": "^3.5.3",
-    "primereact": "^10.9.6",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.56.1",
@@ -73,7 +72,7 @@
     "@eslint/js": "^9.25.1",
     "@playwright/test": "^1.52.0",
     "@react-router/dev": "^7.6.0",
-    "@tailwindcss/vite": "^4.1.8",
+    "@tailwindcss/vite": "^4.0.0",
     "@types/html-to-text": "^9.0.4",
     "@types/node": "^20",
     "@types/nodemailer": "^6.4.17",
@@ -89,7 +88,7 @@
     "kysely-supabase": "^0.2.0",
     "react-router-devtools": "^5.0.5",
     "supabase": "^2.23.4",
-    "tailwindcss": "^4.1.8",
+    "tailwindcss": "^4.0.0",
     "tw-animate-css": "^1.2.8",
     "typescript": "^5.7.2",
     "typescript-eslint": "^8.31.1",
